sub-text/backend/requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-aiplatform
chromadb
pymupdf
python-dotenv
pydantic
google-genai
reportlab
requests